+++ specs/lib/findAccessorMinMaxSpec.js
@@ -37,8 +37,9 @@ const nonContiguousData = [
 function createGltf(elements, byteStride) {
   const buffer = Buffer.from(new Float32Array(elements).buffer);
   const byteLength = buffer.length;
-  const dataUri =
-    "data:application/octet-stream;base64," + buffer.toString("base64");
+  const dataUri = `data:application/octet-stream;base64,${buffer.toString(
+    "base64"
+  )}`;
   const gltf = {
     asset: {
       version: "2.0",